refactor: unify telemetry service
